+++ CHANGELOG.md
@@ -53,6 +53,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (precompile-distribution) [#38](https://github.com/luchenqun/ethos/pull/38) Implement distribution precompile contract.
 - (precompile-gov) [#39](https://github.com/luchenqun/ethos/pull/39) Implement gov precompile contract.
 - (precompile-bank) [#40](https://github.com/luchenqun/ethos/pull/40) Implement bank precompile contract.
+- (revenue) [#48](https://github.com/luchenqun/ethos/pull/48) Remove revenue module.
+
 
 ### API Breaking